Contracts for pending state overlay.
This implements Solidity contracts that we will use with an eth_call
state overlay to instrument the XayaAccounts contract, so that we can
extract the moves generated from pending transactions.

The TrackingAccounts contract will be laid over the real XayaAccounts
contract deployed on-chain, and the CallForwarder contract will be
laid over the sender's EOA address.